Realsense/requirements.txt
opencv-python
openvino-dev
pyrealsense2
websocket-client